split syscall tests (#62)
