+++ crates/blockifier/src/execution/syscalls/syscall_tests/send_message_to_l1.rs
@@ -1,0 +1,54 @@
+use itertools::concat;
+use starknet_api::core::EthAddress;
+use starknet_api::hash::StarkFelt;
+use starknet_api::stark_felt;
+use starknet_api::transaction::{Calldata, L2ToL1Payload};
+use test_case::test_case;
+
+use crate::abi::abi_utils::selector_from_name;
+use crate::context::ChainInfo;
+use crate::execution::call_info::{CallExecution, MessageToL1, OrderedL2ToL1Message};
+use crate::execution::entry_point::CallEntryPoint;
+use crate::execution::sierra_utils::NATIVE_GAS_PLACEHOLDER;
+use crate::test_utils::contracts::FeatureContract;
+use crate::test_utils::initial_test_state::test_state;
+use crate::test_utils::{trivial_external_entry_point_new, CairoVersion, BALANCE};
+
+#[test_case(FeatureContract::SierraTestContract, NATIVE_GAS_PLACEHOLDER; "Native")]
+#[test_case(FeatureContract::TestContract(CairoVersion::Cairo1), 37990; "VM")]
+fn test_send_message_to_l1(test_contract: FeatureContract, expected_gas: u64) {
+    let chain_info = &ChainInfo::create_for_testing();
+    let mut state = test_state(chain_info, BALANCE, &[(test_contract, 1)]);
+
+    let to_address = stark_felt!(1234_u16);
+    let payload = vec![stark_felt!(2019_u16), stark_felt!(2020_u16), stark_felt!(2021_u16)];
+    let calldata = Calldata(
+        concat(vec![
+            vec![
+                to_address,
+                // TODO(Ori, 1/2/2024): Write an indicative expect message explaining why the
+                // convertion works.
+                stark_felt!(u64::try_from(payload.len()).expect("Failed to convert usize to u64.")),
+            ],
+            payload.clone(),
+        ])
+        .into(),
+    );
+    let entry_point_call = CallEntryPoint {
+        entry_point_selector: selector_from_name("test_send_message_to_l1"),
+        calldata,
+        ..trivial_external_entry_point_new(test_contract)
+    };
+
+    let to_address = EthAddress::try_from(to_address).unwrap();
+    let message = MessageToL1 { to_address, payload: L2ToL1Payload(payload) };
+
+    pretty_assertions::assert_eq!(
+        entry_point_call.execute_directly(&mut state).unwrap().execution,
+        CallExecution {
+            l2_to_l1_messages: vec![OrderedL2ToL1Message { order: 0, message }],
+            gas_consumed: expected_gas,
+            ..Default::default()
+        }
+    );
+}
